feat: Sistema completo de gestión - Versión 1.0 estable
Mejoras de rendimiento y optimizaciones:
- Implementación de caché Caffeine con TTL de 5 minutos para reportes
- Endpoint consolidado /dashboard-completo reduce 5 llamadas API a 1
- Invalidación automática de caché en operaciones de ventas/compras

Nuevas funcionalidades:
- Sistema de reportes con dashboard consolidado y gráficos interactivos
- Módulo de compras completo con gestión de proveedores
- Módulo de ventas con POS, múltiples métodos de pago y exportación de tickets
- DTOs para reportes: ventas, compras, inventario, rentabilidad, flujo de caja
- Entidades de métricas: AlertaInventario, MetricaDiaria, HistorialReporte

Mejoras de UI/UX:
- Diseño responsive optimizado (mobile-first: 640px, 768px, 1024px)
- Reemplazo de emojis por iconos SVG de Material Icons
- Fix de overflow que bloqueaba scroll vertical
- Interfaz moderna con animaciones fluidas

Refactorización y limpieza:
- Eliminación de componente Configuracion (no requerido)
- Corrección de errores BigDecimal en ComprasController y VentaService
- Eliminación de scripts SQL duplicados de fix
- Actualización de modelos con validaciones mejoradas

Documentación:
- README simplificado y profesional
- Eliminación de información confidencial
- Documentación actualizada al estado de desarrollo completado
- Guías de instalación y configuración mejoradas

BREAKING CHANGES: Ninguno - Versión estable y retrocompatible

Estado: ✅ Sistema funcional y en producción

+++ src/main/java/com/snayber/api_jdbc/service/VentaService.java
@@ -66,7 +66,8 @@ public class VentaService {
         Producto prod = producto.get();
 
         // Convertir cantidad de venta a BigDecimal para comparar con el stock del producto
-        BigDecimal cantidadVenta = BigDecimal.valueOf(venta.getCantidad());
+    // cantidad ya es BigDecimal en el modelo
+    BigDecimal cantidadVenta = venta.getCantidad();
 
         // Validar stock disponible
         if (prod.getCantidad().compareTo(cantidadVenta) < 0) {
@@ -85,11 +86,11 @@ public class VentaService {
         venta.setFechaVenta(LocalDateTime.now());
 
         // Calcular precio total
-        BigDecimal precioTotal = venta.getPrecioUnitario().multiply(BigDecimal.valueOf(venta.getCantidad()));
+    BigDecimal precioTotal = venta.getPrecioUnitario().multiply(venta.getCantidad());
         venta.setPrecioTotal(precioTotal);
 
         // Calcular ganancia
-        BigDecimal costoTotal = venta.getCostoUnitario().multiply(BigDecimal.valueOf(venta.getCantidad()));
+    BigDecimal costoTotal = venta.getCostoUnitario().multiply(venta.getCantidad());
         BigDecimal ganancia = precioTotal.subtract(costoTotal);
         venta.setGanancia(ganancia);
 
@@ -111,7 +112,7 @@ public class VentaService {
             Optional<Producto> productoOpt = productoRepository.findById(venta.getProductoId());
             if (productoOpt.isPresent()) {
                 Producto producto = productoOpt.get();
-                BigDecimal cantidadVenta = BigDecimal.valueOf(venta.getCantidad());
+                BigDecimal cantidadVenta = venta.getCantidad();
                 producto.setCantidad(producto.getCantidad().add(cantidadVenta));
                 productoRepository.save(producto);
             }
